refector(ParticipantsTest): fixture를 사용하도록 수정
